# src/main/java/com/hobbyhub/hobbyhub/dto/CreateGroupRequest.java
package com.hobbyhub.hobbyhub.dto;

public class CreateGroupRequest {
    private String name;
    private String hobby;
    private String description;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getHobby() {
        return hobby;
    }

    public void setHobby(String hobby) {
        this.hobby = hobby;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}