# src/main/java/com/hobbyhub/hobbyhub/dto/SignInRequest.java
package com.hobbyhub.hobbyhub.dto;

public class SignInRequest {
    private String email;
    private String password;

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}